feat: v1.0.0 — stable public release
Version: bump all crates and HUD to 1.0.0

Core:
- Extract cadis-core modules: tools.rs, voice.rs, orchestrator.rs, workspace.rs
- Telegram adapter: add DaemonBridge for cadisd socket forwarding

HUD:
- Disable send/approve/deny buttons when disconnected
- Add patch.created and test.result event handlers
- Add theme, rename, and voice prefs tests
- Open-source cleanup: remove remaining RamaClaw references

Docs:
- Fix stale known limitations (HUD packaging exists, installer exists)
- Add wiki content: Home, Getting Started, Configuration, FAQ, Troubleshooting
- Add screenshot parity verification script
- Update UI Feature Parity Checklist

302 tests pass. cargo fmt/clippy/build --release clean.
