+++ apps/cadis-hud/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cadis-hud",
-  "version": "0.9.2",
+  "version": "1.0.0",
   "private": true,
   "description": "CADIS Linux desktop multi-agent HUD (Tauri + React)",
   "packageManager": "pnpm@10.33.2",
