+++ crates/cadis-avatar/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cadis-avatar"
-version = "0.9.2"
+version = "1.0.0"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
