+++ crates/cadis-core/src/lib.rs
@@ -50,6 +50,16 @@ use cadis_store::{
 use chrono::{DateTime, Duration, SecondsFormat, Utc};
 use serde::{Deserialize, Serialize};
 
+mod orchestrator;
+mod tools;
+mod voice;
+mod workspace;
+
+use orchestrator::*;
+use tools::*;
+use voice::*;
+use workspace::*;
+
 const FILE_READ_LIMIT_BYTES: usize = 64 * 1024;
 const FILE_SEARCH_LIMIT_BYTES: u64 = 1024 * 1024;
 const FILE_SEARCH_DEFAULT_LIMIT: usize = 50;
@@ -4788,172 +4798,6 @@ impl OrchestratorConfig {
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-struct Orchestrator {
-    config: OrchestratorConfig,
-}
-
-impl Orchestrator {
-    fn new(config: OrchestratorConfig) -> Self {
-        Self { config }
-    }
-
-    fn route_message(
-        &self,
-        explicit_agent_id: Option<AgentId>,
-        content: &str,
-        agents: &HashMap<AgentId, AgentRecord>,
-    ) -> Result<OrchestratorDecision, RouteError> {
-        let route = if let Some(agent_id) = explicit_agent_id {
-            let content = self.strip_matching_leading_mention(&agent_id, content, agents);
-            self.route_to_agent(
-                agent_id,
-                content,
-                "explicit target_agent_id".to_owned(),
-                None,
-                agents,
-            )?
-        } else if let Some((mention, remaining)) = leading_mention(content) {
-            let Some(agent_id) = self.resolve_agent_mention(&mention, agents) else {
-                return Err(RouteError {
-                    code: "agent_not_found",
-                    message: format!("no agent matches @{mention}"),
-                });
-            };
-            self.route_to_agent(
-                agent_id,
-                remaining,
-                format!("@{mention} mention"),
-                None,
-                agents,
-            )?
-        } else {
-            self.route_to_agent(
-                AgentId::from("main"),
-                content.to_owned(),
-                "default orchestrator".to_owned(),
-                None,
-                agents,
-            )?
-        };
-
-        if route.agent_id.as_str() == "main" {
-            self.apply_explicit_action(route, agents)
-        } else {
-            Ok(OrchestratorDecision::Route(route))
-        }
-    }
-
-    fn apply_explicit_action(
-        &self,
-        route: RouteDecision,
-        agents: &HashMap<AgentId, AgentRecord>,
-    ) -> Result<OrchestratorDecision, RouteError> {
-        let Some(action) = parse_orchestrator_action(&route.content, &self.config) else {
-            return Ok(OrchestratorDecision::Route(route));
-        };
-
-        if !self.config.worker_delegation_enabled {
-            return Err(RouteError {
-                code: "orchestrator_worker_delegation_disabled",
-                message: "orchestrator worker delegation actions are disabled".to_owned(),
-            });
-        }
-
-        match action {
-            ExplicitOrchestratorAction::Route {
-                mention,
-                content,
-                worker_summary,
-            } => {
-                let Some(agent_id) = self.resolve_agent_mention(&mention, agents) else {
-                    return Err(RouteError {
-                        code: "agent_not_found",
-                        message: format!("no agent matches @{mention}"),
-                    });
-                };
-                self.route_to_agent(
-                    agent_id,
-                    content,
-                    format!("orchestrator action: route @{mention}"),
-                    Some(worker_summary),
-                    agents,
-                )
-                .map(OrchestratorDecision::Route)
-            }
-            ExplicitOrchestratorAction::Spawn {
-                role,
-                content,
-                worker_summary,
-            } => Ok(OrchestratorDecision::SpawnAndRoute(SpawnRouteDecision {
-                role,
-                content,
-                reason: "orchestrator action: spawn worker".to_owned(),
-                worker_summary,
-            })),
-        }
-    }
-
-    fn route_to_agent(
-        &self,
-        agent_id: AgentId,
-        content: String,
-        reason: String,
-        worker_summary: Option<String>,
-        agents: &HashMap<AgentId, AgentRecord>,
-    ) -> Result<RouteDecision, RouteError> {
-        let Some(agent) = agents.get(&agent_id) else {
-            return Err(RouteError {
-                code: "agent_not_found",
-                message: format!("agent '{agent_id}' was not found"),
-            });
-        };
-        Ok(RouteDecision {
-            agent_id: agent.id.clone(),
-            agent_name: agent.display_name.clone(),
-            content: normalize_route_content(content),
-            reason,
-            worker_summary,
-        })
-    }
-
-    fn resolve_agent_mention(
-        &self,
-        mention: &str,
-        agents: &HashMap<AgentId, AgentRecord>,
-    ) -> Option<AgentId> {
-        let normalized = normalize_lookup(mention);
-        agents
-            .values()
-            .find(|agent| {
-                [
-                    agent.id.as_str(),
-                    agent.display_name.as_str(),
-                    agent.role.as_str(),
-                ]
-                .into_iter()
-                .any(|candidate| normalize_lookup(candidate) == normalized)
-            })
-            .map(|agent| agent.id.clone())
-    }
-
-    fn strip_matching_leading_mention(
-        &self,
-        agent_id: &AgentId,
-        content: &str,
-        agents: &HashMap<AgentId, AgentRecord>,
-    ) -> String {
-        let Some((mention, remaining)) = leading_mention(content) else {
-            return content.to_owned();
-        };
-        if self.resolve_agent_mention(&mention, agents).as_ref() == Some(agent_id) {
-            remaining
-        } else {
-            content.to_owned()
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
 struct SessionRecord {
     title: Option<String>,
     _cwd: Option<String>,
@@ -5221,150 +5065,6 @@ impl AgentSessionRecord {
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-struct WorkspaceRecord {
-    id: WorkspaceId,
-    kind: WorkspaceKind,
-    root: PathBuf,
-    aliases: Vec<String>,
-    vcs: Option<String>,
-    trusted: bool,
-    worktree_root: Option<String>,
-    artifact_root: Option<String>,
-}
-
-impl WorkspaceRecord {
-    fn event_payload(self) -> WorkspaceRecordPayload {
-        WorkspaceRecordPayload {
-            workspace_id: self.id,
-            kind: self.kind,
-            root: self.root.display().to_string(),
-            aliases: self.aliases,
-            vcs: self.vcs,
-            trusted: self.trusted,
-            worktree_root: self.worktree_root,
-            artifact_root: self.artifact_root,
-        }
-    }
-
-    fn into_store(self) -> WorkspaceMetadata {
-        WorkspaceMetadata {
-            id: self.id.to_string(),
-            kind: store_workspace_kind(self.kind),
-            root: self.root,
-            vcs: store_workspace_vcs(self.vcs.as_deref()),
-            owner: None,
-            trusted: self.trusted,
-            worktree_root: self.worktree_root.map(PathBuf::from),
-            artifact_root: self.artifact_root.map(PathBuf::from),
-            checkpoint_policy: CheckpointPolicy::Disabled,
-            aliases: if self.aliases.is_empty() {
-                Vec::new()
-            } else {
-                vec![WorkspaceAlias {
-                    workspace_id: self.id.to_string(),
-                    aliases: self.aliases,
-                }]
-            },
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct WorkspaceGrantRecord {
-    grant_id: WorkspaceGrantId,
-    agent_id: Option<AgentId>,
-    workspace_id: WorkspaceId,
-    root: PathBuf,
-    access: Vec<WorkspaceAccess>,
-    created_at: Timestamp,
-    expires_at: Option<Timestamp>,
-    source: String,
-}
-
-impl WorkspaceGrantRecord {
-    fn event_payload(self) -> WorkspaceGrantPayload {
-        WorkspaceGrantPayload {
-            grant_id: self.grant_id,
-            agent_id: self.agent_id,
-            workspace_id: self.workspace_id,
-            root: self.root.display().to_string(),
-            access: self.access,
-            expires_at: self.expires_at,
-            source: self.source,
-        }
-    }
-
-    fn into_store(self, profile_id: &str) -> StoreWorkspaceGrantRecord {
-        StoreWorkspaceGrantRecord {
-            grant_id: self.grant_id.to_string(),
-            profile_id: profile_id.to_owned(),
-            agent_id: self.agent_id,
-            workspace_id: self.workspace_id.to_string(),
-            root: self.root,
-            access: self
-                .access
-                .into_iter()
-                .map(store_workspace_access)
-                .collect(),
-            created_at: self.created_at,
-            expires_at: self.expires_at,
-            source: store_grant_source(&self.source),
-            reason: None,
-        }
-    }
-
-    fn is_expired(&self) -> bool {
-        self.expires_at
-            .as_ref()
-            .and_then(|expires_at| DateTime::parse_from_rfc3339(expires_at.as_str()).ok())
-            .map(|expires_at| expires_at.with_timezone(&Utc) <= Utc::now())
-            .unwrap_or(false)
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ResolvedWorkspace {
-    root: PathBuf,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct RouteDecision {
-    agent_id: AgentId,
-    agent_name: String,
-    content: String,
-    reason: String,
-    worker_summary: Option<String>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct SpawnRouteDecision {
-    role: String,
-    content: String,
-    reason: String,
-    worker_summary: String,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-enum OrchestratorDecision {
-    Route(RouteDecision),
-    SpawnAndRoute(SpawnRouteDecision),
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-enum ExplicitOrchestratorAction {
-    Route {
-        mention: String,
-        content: String,
-        worker_summary: String,
-    },
-    Spawn {
-        role: String,
-        content: String,
-        worker_summary: String,
-    },
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
 struct WorkerDelegation {
     worker_id: String,
     agent_session_id: AgentSessionId,
@@ -5550,874 +5250,6 @@ impl WorkerMetadata {
 struct PendingApproval {
     record: ApprovalRecord,
     request: Option<ToolCallRequest>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct VoicePreflightRecord {
-    surface: String,
-    status: String,
-    summary: String,
-    checked_at: Timestamp,
-    checks: Vec<VoiceDoctorCheck>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct VoiceRuntimePreferences {
-    enabled: bool,
-    provider: String,
-    voice_id: String,
-    stt_language: String,
-    rate: i16,
-    pitch: i16,
-    volume: i16,
-    auto_speak: bool,
-    max_spoken_chars: usize,
-}
-
-impl VoiceRuntimePreferences {
-    fn from_options(options: &serde_json::Value) -> Self {
-        let voice = options.get("voice").and_then(serde_json::Value::as_object);
-
-        Self {
-            enabled: voice
-                .and_then(|voice| voice.get("enabled"))
-                .and_then(serde_json::Value::as_bool)
-                .unwrap_or(false),
-            provider: voice
-                .and_then(|voice| voice.get("provider"))
-                .and_then(serde_json::Value::as_str)
-                .map(normalize_voice_config_string)
-                .filter(|value| !value.is_empty())
-                .unwrap_or_else(|| "edge".to_owned()),
-            voice_id: voice
-                .and_then(|voice| voice.get("voice_id"))
-                .and_then(serde_json::Value::as_str)
-                .map(normalize_voice_config_string)
-                .unwrap_or_else(|| "id-ID-GadisNeural".to_owned()),
-            stt_language: voice
-                .and_then(|voice| voice.get("stt_language"))
-                .and_then(serde_json::Value::as_str)
-                .map(normalize_voice_config_string)
-                .filter(|value| !value.is_empty())
-                .unwrap_or_else(|| "auto".to_owned()),
-            rate: voice
-                .and_then(|voice| voice.get("rate"))
-                .and_then(serde_json::Value::as_i64)
-                .and_then(|value| i16::try_from(value).ok())
-                .map(clamp_voice_adjustment)
-                .unwrap_or_default(),
-            pitch: voice
-                .and_then(|voice| voice.get("pitch"))
-                .and_then(serde_json::Value::as_i64)
-                .and_then(|value| i16::try_from(value).ok())
-                .map(clamp_voice_adjustment)
-                .unwrap_or_default(),
-            volume: voice
-                .and_then(|voice| voice.get("volume"))
-                .and_then(serde_json::Value::as_i64)
-                .and_then(|value| i16::try_from(value).ok())
-                .map(clamp_voice_adjustment)
-                .unwrap_or_default(),
-            auto_speak: voice
-                .and_then(|voice| voice.get("auto_speak"))
-                .and_then(serde_json::Value::as_bool)
-                .unwrap_or(false),
-            max_spoken_chars: voice
-                .and_then(|voice| voice.get("max_spoken_chars"))
-                .and_then(serde_json::Value::as_u64)
-                .and_then(|value| usize::try_from(value).ok())
-                .filter(|value| *value > 0)
-                .unwrap_or(800),
-        }
-    }
-
-    fn from_preview(options: &serde_json::Value, prefs: Option<VoicePreferences>) -> Self {
-        let mut runtime_prefs = Self::from_options(options);
-        if let Some(prefs) = prefs {
-            runtime_prefs.voice_id = normalize_voice_config_string(&prefs.voice_id);
-            runtime_prefs.rate = clamp_voice_adjustment(prefs.rate);
-            runtime_prefs.pitch = clamp_voice_adjustment(prefs.pitch);
-            runtime_prefs.volume = clamp_voice_adjustment(prefs.volume);
-        }
-        runtime_prefs
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum TtsProviderKind {
-    Edge,
-    OpenAi,
-    System,
-    Stub,
-    Unsupported,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct StubbedTtsProvider {
-    kind: TtsProviderKind,
-    configured_id: String,
-}
-
-impl StubbedTtsProvider {
-    fn new(provider: &str) -> Self {
-        Self {
-            kind: tts_provider_kind(provider),
-            configured_id: normalize_voice_config_string(provider),
-        }
-    }
-}
-
-impl TtsProvider for StubbedTtsProvider {
-    fn id(&self) -> &'static str {
-        match self.kind {
-            TtsProviderKind::Edge => "edge",
-            TtsProviderKind::OpenAi => "openai",
-            TtsProviderKind::System => "system",
-            TtsProviderKind::Stub => "stub",
-            TtsProviderKind::Unsupported => "unsupported",
-        }
-    }
-
-    fn label(&self) -> &'static str {
-        match self.kind {
-            TtsProviderKind::Edge => "Edge TTS daemon stub",
-            TtsProviderKind::OpenAi => "OpenAI TTS daemon stub",
-            TtsProviderKind::System => "System speech daemon stub",
-            TtsProviderKind::Stub => "Deterministic test TTS stub",
-            TtsProviderKind::Unsupported => "Unsupported TTS provider",
-        }
-    }
-
-    fn supported_voices(&self) -> Vec<TtsVoice> {
-        curated_tts_voices()
-    }
-
-    fn speak(&mut self, request: TtsRequest<'_>) -> Result<TtsOutput, TtsError> {
-        if self.kind == TtsProviderKind::Unsupported {
-            return Err(TtsError::new(
-                "unsupported_tts_provider",
-                format!("unsupported TTS provider '{}'", self.configured_id),
-                false,
-            ));
-        }
-        Ok(TtsOutput {
-            provider: self.id().to_owned(),
-            voice_id: request.voice_id.to_owned(),
-            spoken_chars: request.text.chars().count(),
-            audio_path: None,
-        })
-    }
-
-    fn stop(&mut self) -> Result<(), TtsError> {
-        if self.kind == TtsProviderKind::Unsupported {
-            return Err(TtsError::new(
-                "unsupported_tts_provider",
-                format!("unsupported TTS provider '{}'", self.configured_id),
-                false,
-            ));
-        }
-        Ok(())
-    }
-}
-
-/// Daemon-owned Edge TTS provider that calls the `edge-tts` Python CLI.
-struct EdgeTtsProvider;
-
-impl TtsProvider for EdgeTtsProvider {
-    fn id(&self) -> &'static str {
-        "edge"
-    }
-
-    fn label(&self) -> &'static str {
-        "Edge TTS (daemon subprocess)"
-    }
-
-    fn supported_voices(&self) -> Vec<TtsVoice> {
-        curated_tts_voices()
-    }
-
-    fn speak(&mut self, request: TtsRequest<'_>) -> Result<TtsOutput, TtsError> {
-        // Validate voice_id: only allow alphanumeric, dash, and dot.
-        if request.voice_id.is_empty()
-            || !request
-                .voice_id
-                .chars()
-                .all(|c| c.is_alphanumeric() || c == '-' || c == '.')
-        {
-            return Err(TtsError::new(
-                "invalid_voice_id",
-                "voice_id must be non-empty and contain only alphanumeric, dash, or dot characters",
-                false,
-            ));
-        }
-
-        // Truncate text to prevent excessively long subprocess arguments.
-        const MAX_TTS_TEXT_CHARS: usize = 5000;
-        let text = if request.text.chars().count() > MAX_TTS_TEXT_CHARS {
-            truncate_to_utf8_boundary(request.text, MAX_TTS_TEXT_CHARS).0
-        } else {
-            request.text
-        };
-
-        let temp_dir = std::env::temp_dir().join("cadis-edge-tts");
-        let _ = fs::create_dir_all(&temp_dir);
-        // Use PID + nanos for a unique path per invocation.
-        let nanos = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_nanos();
-        let audio_path = temp_dir.join(format!("cadis-tts-{}-{nanos}.mp3", std::process::id()));
-
-        let rate_arg = format!("{:+}%", request.rate);
-        let pitch_arg = format!("{:+}Hz", request.pitch);
-        let volume_arg = format!("{:+}%", request.volume);
-
-        let output = Command::new("edge-tts")
-            .args(["--voice", request.voice_id])
-            .args(["--rate", &rate_arg])
-            .args(["--pitch", &pitch_arg])
-            .args(["--volume", &volume_arg])
-            .args(["--text", text])
-            .arg("--write-media")
-            .arg(&audio_path)
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .stderr(Stdio::piped())
-            .output();
-
-        match output {
-            Ok(result) if result.status.success() => Ok(TtsOutput {
-                provider: "edge".to_owned(),
-                voice_id: request.voice_id.to_owned(),
-                spoken_chars: request.text.chars().count(),
-                audio_path: Some(audio_path),
-            }),
-            Ok(result) => {
-                let stderr = String::from_utf8_lossy(&result.stderr);
-                Err(TtsError::new(
-                    "edge_tts_failed",
-                    format!(
-                        "edge-tts exited with code {:?}: {}",
-                        result.status.code(),
-                        stderr.trim()
-                    ),
-                    true,
-                ))
-            }
-            Err(error) if error.kind() == io::ErrorKind::NotFound => Err(TtsError::new(
-                "edge_tts_not_found",
-                "edge-tts binary is not installed; install with: pip install edge-tts",
-                false,
-            )),
-            Err(error) => Err(TtsError::new(
-                "edge_tts_spawn_failed",
-                error.to_string(),
-                true,
-            )),
-        }
-    }
-
-    fn stop(&mut self) -> Result<(), TtsError> {
-        Ok(())
-    }
-}
-
-fn tts_provider_from_config(provider: &str) -> Box<dyn TtsProvider> {
-    match provider {
-        "edge" => Box::new(EdgeTtsProvider),
-        _ => Box::new(StubbedTtsProvider::new(provider)),
-    }
-}
-
-fn tts_provider_kind(provider: &str) -> TtsProviderKind {
-    match provider {
-        "edge" => TtsProviderKind::Edge,
-        "openai" => TtsProviderKind::OpenAi,
-        "system" => TtsProviderKind::System,
-        "stub" => TtsProviderKind::Stub,
-        _ => TtsProviderKind::Unsupported,
-    }
-}
-
-fn curated_tts_voices() -> Vec<TtsVoice> {
-    vec![
-        TtsVoice {
-            id: "id-ID-ArdiNeural",
-            label: "Ardi (Indonesian, Male)",
-            locale: "id-ID",
-            gender: "Male",
-        },
-        TtsVoice {
-            id: "id-ID-GadisNeural",
-            label: "Gadis (Indonesian, Female)",
-            locale: "id-ID",
-            gender: "Female",
-        },
-        TtsVoice {
-            id: "ms-MY-OsmanNeural",
-            label: "Osman (Malay, Male)",
-            locale: "ms-MY",
-            gender: "Male",
-        },
-        TtsVoice {
-            id: "ms-MY-YasminNeural",
-            label: "Yasmin (Malay, Female)",
-            locale: "ms-MY",
-            gender: "Female",
-        },
-        TtsVoice {
-            id: "en-US-AvaNeural",
-            label: "Ava (US, Female)",
-            locale: "en-US",
-            gender: "Female",
-        },
-        TtsVoice {
-            id: "en-US-AndrewNeural",
-            label: "Andrew (US, Male)",
-            locale: "en-US",
-            gender: "Male",
-        },
-        TtsVoice {
-            id: "en-US-EmmaNeural",
-            label: "Emma (US, Female)",
-            locale: "en-US",
-            gender: "Female",
-        },
-        TtsVoice {
-            id: "en-US-BrianNeural",
-            label: "Brian (US, Male)",
-            locale: "en-US",
-            gender: "Male",
-        },
-        TtsVoice {
-            id: "en-GB-SoniaNeural",
-            label: "Sonia (GB, Female)",
-            locale: "en-GB",
-            gender: "Female",
-        },
-        TtsVoice {
-            id: "en-GB-RyanNeural",
-            label: "Ryan (GB, Male)",
-            locale: "en-GB",
-            gender: "Male",
-        },
-    ]
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum SpeechMode {
-    AutoSpeak,
-    Preview,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-enum SpeechDecision {
-    Speak,
-    Blocked(&'static str),
-    RequiresSummary(&'static str),
-}
-
-fn speech_decision(
-    prefs: &VoiceRuntimePreferences,
-    content_kind: ContentKind,
-    text: &str,
-    mode: SpeechMode,
-) -> SpeechDecision {
-    let text = text.trim();
-    if text.is_empty() {
-        return SpeechDecision::Blocked("empty_text");
-    }
-
-    if mode == SpeechMode::AutoSpeak {
-        if !prefs.enabled {
-            return SpeechDecision::Blocked("voice_disabled");
-        }
-        if !prefs.auto_speak {
-            return SpeechDecision::Blocked("auto_speak_disabled");
-        }
-    }
-
-    match content_kind {
-        ContentKind::Code => return SpeechDecision::Blocked("code_not_speakable"),
-        ContentKind::Diff => return SpeechDecision::Blocked("diff_not_speakable"),
-        ContentKind::TerminalLog => return SpeechDecision::Blocked("terminal_log_not_speakable"),
-        ContentKind::TestResult if text.chars().count() > prefs.max_spoken_chars => {
-            return SpeechDecision::Blocked("long_tool_output_not_speakable");
-        }
-        ContentKind::TestResult if looks_like_raw_tool_output(text) => {
-            return SpeechDecision::Blocked("raw_tool_output_not_speakable");
-        }
-        _ => {}
-    }
-
-    if text.chars().count() > prefs.max_spoken_chars {
-        return SpeechDecision::RequiresSummary("content_exceeds_max_spoken_chars");
-    }
-
-    SpeechDecision::Speak
-}
-
-/// Truncates text to the first 2-3 sentences for spoken summary.
-fn summarize_for_speech(text: &str) -> String {
-    let mut end = 0;
-    let mut sentences = 0;
-    for (index, character) in text.char_indices() {
-        if matches!(character, '.' | '!' | '?') {
-            let next = text[index + character.len_utf8()..].chars().next();
-            if next.is_none() || next.is_some_and(|c| c.is_whitespace()) {
-                sentences += 1;
-                end = index + character.len_utf8();
-                if sentences >= 3 {
-                    break;
-                }
-            }
-        }
-    }
-    if end == 0 {
-        text.split_whitespace()
-            .take(30)
-            .collect::<Vec<_>>()
-            .join(" ")
-    } else {
-        text[..end].trim().to_owned()
-    }
-}
-
-/// Generates a short spoken risk summary for an approval request.
-fn approval_risk_speech(record: &ApprovalRecord) -> String {
-    let mut speech = format!("Approval needed: {}", record.tool_name);
-    if let Some(command) = &record.command {
-        let short = command.chars().take(60).collect::<String>();
-        speech.push_str(&format!(", {short}"));
-    }
-    if let Some(workspace) = &record.workspace {
-        let name = Path::new(workspace)
-            .file_name()
-            .and_then(|n| n.to_str())
-            .unwrap_or(workspace);
-        speech.push_str(&format!(" in workspace {name}"));
-    }
-    speech
-}
-
-fn looks_like_raw_tool_output(text: &str) -> bool {
-    let line_count = text.lines().count();
-    line_count > 12
-        || text.contains("```")
-        || text.contains("diff --git")
-        || text.contains("thread '")
-        || text.contains("panicked at")
-        || text.contains("error[E")
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ToolRegistry {
-    definitions: Vec<ToolDefinition>,
-}
-
-impl ToolRegistry {
-    fn new(definitions: Vec<ToolDefinition>) -> Result<Self, RuntimeError> {
-        for (index, definition) in definitions.iter().enumerate() {
-            if definitions[..index]
-                .iter()
-                .any(|previous| previous.name == definition.name)
-            {
-                return Err(RuntimeError {
-                    code: "duplicate_tool_name",
-                    message: format!("tool '{}' is registered more than once", definition.name),
-                });
-            }
-            if definition.description.trim().is_empty() {
-                return Err(RuntimeError {
-                    code: "invalid_tool_description",
-                    message: format!("tool '{}' is missing a description", definition.name),
-                });
-            }
-            if definition.side_effects.is_empty() {
-                return Err(RuntimeError {
-                    code: "invalid_tool_side_effects",
-                    message: format!(
-                        "tool '{}' must declare at least one side effect",
-                        definition.name
-                    ),
-                });
-            }
-            if definition.timeout_secs == 0 {
-                return Err(RuntimeError {
-                    code: "invalid_tool_timeout",
-                    message: format!("tool '{}' must declare a positive timeout", definition.name),
-                });
-            }
-        }
-
-        Ok(Self { definitions })
-    }
-
-    fn builtin() -> Result<Self, RuntimeError> {
-        Self::new(vec![
-            ToolDefinition::safe_read(
-                "file.read",
-                "Read one file inside an approved workspace",
-                ToolInputSchema::FileRead,
-                &[ToolSideEffect::ReadFiles],
-                5,
-                ToolWorkspaceBehavior::PathScoped,
-            ),
-            ToolDefinition::safe_read(
-                "file.search",
-                "Search text files inside an approved workspace",
-                ToolInputSchema::FileSearch,
-                &[ToolSideEffect::SearchFiles],
-                10,
-                ToolWorkspaceBehavior::PathScoped,
-            ),
-            ToolDefinition::safe_read(
-                "git.status",
-                "Read git status inside an approved workspace",
-                ToolInputSchema::GitStatus,
-                &[ToolSideEffect::ReadGitMetadata],
-                10,
-                ToolWorkspaceBehavior::PathScoped,
-            ),
-            ToolDefinition::approval_placeholder(
-                "file.write",
-                "Write or replace files inside an approved workspace",
-                cadis_protocol::RiskClass::WorkspaceEdit,
-                ToolInputSchema::WorkspaceMutation,
-                &[ToolSideEffect::EditWorkspace],
-                30,
-                ToolCancellationBehavior::Cooperative,
-                ToolWorkspaceBehavior::PathScoped,
-                false,
-                false,
-            ),
-            ToolDefinition::approval_placeholder(
-                "file.patch",
-                "Apply a patch inside an approved workspace",
-                cadis_protocol::RiskClass::WorkspaceEdit,
-                ToolInputSchema::WorkspaceMutation,
-                &[ToolSideEffect::EditWorkspace],
-                30,
-                ToolCancellationBehavior::Cooperative,
-                ToolWorkspaceBehavior::PathScoped,
-                false,
-                false,
-            ),
-            ToolDefinition::safe_read(
-                "git.diff",
-                "Read git diff output for an approved workspace",
-                ToolInputSchema::GitDiff,
-                &[ToolSideEffect::ReadGitDiff],
-                20,
-                ToolWorkspaceBehavior::PathScoped,
-            ),
-            ToolDefinition::approval_placeholder(
-                "git.worktree.create",
-                "Create a CADIS-managed git worktree",
-                cadis_protocol::RiskClass::WorkspaceEdit,
-                ToolInputSchema::GitMutation,
-                &[ToolSideEffect::CreateWorktree],
-                60,
-                ToolCancellationBehavior::Cooperative,
-                ToolWorkspaceBehavior::CadisManagedWorktree,
-                false,
-                false,
-            ),
-            ToolDefinition::approval_placeholder(
-                "git.worktree.remove",
-                "Remove a CADIS-managed git worktree",
-                cadis_protocol::RiskClass::WorkspaceEdit,
-                ToolInputSchema::GitMutation,
-                &[ToolSideEffect::RemoveWorktree],
-                60,
-                ToolCancellationBehavior::Cooperative,
-                ToolWorkspaceBehavior::CadisManagedWorktree,
-                false,
-                false,
-            ),
-            ToolDefinition::approval_placeholder(
-                "shell.run",
-                "Run a local shell command in an approved workspace",
-                cadis_protocol::RiskClass::SystemChange,
-                ToolInputSchema::ShellRun,
-                &[ToolSideEffect::RunSubprocess],
-                900,
-                ToolCancellationBehavior::Cooperative,
-                ToolWorkspaceBehavior::RequiresWorkspace,
-                false,
-                true,
-            ),
-        ])
-    }
-
-    fn get(&self, name: &str) -> Option<&ToolDefinition> {
-        self.definitions
-            .iter()
-            .find(|definition| definition.name == name)
-    }
-
-    fn is_auto_executable_safe_read(&self, name: &str) -> bool {
-        self.get(name)
-            .is_some_and(|definition| definition.execution == ToolExecutionMode::AutoExecute)
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ToolDefinition {
-    name: &'static str,
-    description: &'static str,
-    risk_class: cadis_protocol::RiskClass,
-    input_schema: ToolInputSchema,
-    execution: ToolExecutionMode,
-    side_effects: &'static [ToolSideEffect],
-    timeout_secs: u64,
-    timeout_behavior: ToolTimeoutBehavior,
-    cancellation_behavior: ToolCancellationBehavior,
-    workspace_behavior: ToolWorkspaceBehavior,
-    needs_network: bool,
-    may_read_secrets: bool,
-}
-
-impl ToolDefinition {
-    fn safe_read(
-        name: &'static str,
-        description: &'static str,
-        input_schema: ToolInputSchema,
-        side_effects: &'static [ToolSideEffect],
-        timeout_secs: u64,
-        workspace_behavior: ToolWorkspaceBehavior,
-    ) -> Self {
-        Self {
-            name,
-            description,
-            risk_class: cadis_protocol::RiskClass::SafeRead,
-            input_schema,
-            execution: ToolExecutionMode::AutoExecute,
-            side_effects,
-            timeout_secs,
-            timeout_behavior: ToolTimeoutBehavior::FailClosed,
-            cancellation_behavior: ToolCancellationBehavior::NotSupported,
-            workspace_behavior,
-            needs_network: false,
-            may_read_secrets: false,
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn approval_placeholder(
-        name: &'static str,
-        description: &'static str,
-        risk_class: cadis_protocol::RiskClass,
-        input_schema: ToolInputSchema,
-        side_effects: &'static [ToolSideEffect],
-        timeout_secs: u64,
-        cancellation_behavior: ToolCancellationBehavior,
-        workspace_behavior: ToolWorkspaceBehavior,
-        needs_network: bool,
-        may_read_secrets: bool,
-    ) -> Self {
-        Self {
-            name,
-            description,
-            risk_class,
-            input_schema,
-            execution: ToolExecutionMode::ApprovalPlaceholder,
-            side_effects,
-            timeout_secs,
-            timeout_behavior: ToolTimeoutBehavior::FailClosed,
-            cancellation_behavior,
-            workspace_behavior,
-            needs_network,
-            may_read_secrets,
-        }
-    }
-
-    fn policy_reason(&self) -> String {
-        match self.execution {
-            ToolExecutionMode::AutoExecute => format!(
-                "{}: {} | schema={:?} | timeout={}s | workspace={:?}",
-                self.name,
-                self.description,
-                self.input_schema,
-                self.timeout_secs,
-                self.workspace_behavior,
-            ),
-            ToolExecutionMode::ApprovalPlaceholder => format!(
-                "{}: {} | risk={:?} | schema={:?} | timeout={}s | workspace={:?}",
-                self.name,
-                self.description,
-                self.risk_class,
-                self.input_schema,
-                self.timeout_secs,
-                self.workspace_behavior,
-            ),
-        }
-    }
-
-    fn approval_summary(&self) -> String {
-        format!(
-            "{} requires approval before execution; side effects: {}; cancellation: {:?}; network: {}; secrets: {}",
-            self.name,
-            self.side_effects
-                .iter()
-                .map(|effect| effect.label())
-                .collect::<Vec<_>>()
-                .join(", "),
-            self.cancellation_behavior,
-            self.needs_network,
-            self.may_read_secrets,
-        )
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolInputSchema {
-    FileRead,
-    FileSearch,
-    GitStatus,
-    GitDiff,
-    ShellRun,
-    WorkspaceMutation,
-    GitMutation,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolExecutionMode {
-    AutoExecute,
-    ApprovalPlaceholder,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolSideEffect {
-    ReadFiles,
-    SearchFiles,
-    ReadGitMetadata,
-    EditWorkspace,
-    ReadGitDiff,
-    CreateWorktree,
-    RemoveWorktree,
-    RunSubprocess,
-}
-
-impl ToolSideEffect {
-    fn label(self) -> &'static str {
-        match self {
-            Self::ReadFiles => "read_files",
-            Self::SearchFiles => "search_files",
-            Self::ReadGitMetadata => "read_git_metadata",
-            Self::EditWorkspace => "edit_workspace",
-            Self::ReadGitDiff => "read_git_diff",
-            Self::CreateWorktree => "create_worktree",
-            Self::RemoveWorktree => "remove_worktree",
-            Self::RunSubprocess => "run_subprocess",
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolTimeoutBehavior {
-    FailClosed,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolCancellationBehavior {
-    NotSupported,
-    Cooperative,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ToolWorkspaceBehavior {
-    PathScoped,
-    RequiresWorkspace,
-    CadisManagedWorktree,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-struct ToolExecutionResult {
-    summary: String,
-    output: serde_json::Value,
-}
-
-// ── Track D: ToolExecutor trait ──────────────────────────────────────
-// Defined as a planned extension point for structured tool backends.
-// Existing tools dispatch through match arms; migration to this trait
-// is tracked as future Track D work.
-
-/// Context passed to tool executors at invocation time.
-#[allow(dead_code)]
-struct ToolContext {
-    workspace: PathBuf,
-    tool_name: String,
-    timeout_secs: u64,
-}
-
-/// Trait for structured tool execution backends.
-#[allow(dead_code)]
-trait ToolExecutor: Send + Sync {
-    fn execute(
-        &self,
-        input: &serde_json::Value,
-        context: &ToolContext,
-    ) -> Result<ToolExecutionResult, RuntimeError>;
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-enum FilePatchOperation {
-    Write {
-        path: String,
-        content: String,
-    },
-    Replace {
-        path: String,
-        old: String,
-        new: String,
-    },
-}
-
-impl FilePatchOperation {
-    fn path(&self) -> &str {
-        match self {
-            Self::Write { path, .. } | Self::Replace { path, .. } => path,
-        }
-    }
-
-    fn action(&self) -> &'static str {
-        match self {
-            Self::Write { .. } => "write",
-            Self::Replace { .. } => "replace",
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct PreparedFilePatch {
-    path: PathBuf,
-    display_path: String,
-    action: &'static str,
-    content: String,
-    mtime: Option<std::time::SystemTime>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct BoundedOutput {
-    bytes: Vec<u8>,
-    truncated: bool,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ShellRunResult {
-    status_success: bool,
-    exit_code: Option<i32>,
-    stdout: BoundedOutput,
-    stderr: BoundedOutput,
-    timed_out: bool,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct SearchMatch {
-    path: String,
-    line_number: usize,
-    line: String,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -6882,112 +5714,6 @@ fn resolve_project_path(project_root: &Path, path: impl AsRef<Path>) -> PathBuf 
     }
 }
 
-fn load_workspace_registry(profile_home: &ProfileHome) -> HashMap<WorkspaceId, WorkspaceRecord> {
-    profile_home
-        .workspace_registry()
-        .load()
-        .unwrap_or_default()
-        .workspace
-        .into_iter()
-        .filter_map(workspace_record_from_store)
-        .map(|record| (record.id.clone(), record))
-        .collect()
-}
-
-fn save_workspace_registry(
-    profile_home: &ProfileHome,
-    workspaces: &HashMap<WorkspaceId, WorkspaceRecord>,
-) -> Result<(), cadis_store::StoreError> {
-    let mut records = workspaces.values().cloned().collect::<Vec<_>>();
-    records.sort_by(|left, right| left.id.cmp(&right.id));
-    profile_home.workspace_registry().save(&WorkspaceRegistry {
-        workspace: records
-            .into_iter()
-            .map(WorkspaceRecord::into_store)
-            .collect(),
-    })
-}
-
-fn load_workspace_grants(
-    profile_home: &ProfileHome,
-    workspaces: &HashMap<WorkspaceId, WorkspaceRecord>,
-) -> HashMap<WorkspaceGrantId, WorkspaceGrantRecord> {
-    profile_home
-        .workspace_grants()
-        .load()
-        .map(|recovery| recovery.records)
-        .unwrap_or_default()
-        .into_iter()
-        .filter_map(|record| workspace_grant_record_from_store(record, workspaces))
-        .map(|record| (record.grant_id.clone(), record))
-        .collect()
-}
-
-fn workspace_record_from_store(record: WorkspaceMetadata) -> Option<WorkspaceRecord> {
-    let root = record.root.canonicalize().ok()?;
-    let aliases = record
-        .aliases
-        .into_iter()
-        .filter(|alias| alias.workspace_id == record.id)
-        .flat_map(|alias| alias.aliases)
-        .collect::<Vec<_>>();
-    Some(WorkspaceRecord {
-        id: WorkspaceId::from(record.id),
-        kind: protocol_workspace_kind(record.kind),
-        root,
-        aliases: normalize_aliases(aliases),
-        vcs: match record.vcs {
-            WorkspaceVcs::Git => Some("git".to_owned()),
-            WorkspaceVcs::None => None,
-        },
-        trusted: record.trusted,
-        worktree_root: record
-            .worktree_root
-            .map(|path| path.display().to_string())
-            .filter(|value| !value.trim().is_empty()),
-        artifact_root: record
-            .artifact_root
-            .map(|path| path.display().to_string())
-            .filter(|value| !value.trim().is_empty()),
-    })
-}
-
-fn workspace_grant_record_from_store(
-    record: StoreWorkspaceGrantRecord,
-    workspaces: &HashMap<WorkspaceId, WorkspaceRecord>,
-) -> Option<WorkspaceGrantRecord> {
-    let workspace_id = WorkspaceId::from(record.workspace_id);
-    if !workspaces.contains_key(&workspace_id) {
-        return None;
-    }
-    Some(WorkspaceGrantRecord {
-        grant_id: WorkspaceGrantId::from(record.grant_id),
-        agent_id: record.agent_id,
-        workspace_id,
-        root: record.root.canonicalize().unwrap_or(record.root),
-        access: normalize_workspace_access(
-            record
-                .access
-                .into_iter()
-                .map(protocol_workspace_access)
-                .collect(),
-        ),
-        created_at: record.created_at,
-        expires_at: record.expires_at,
-        source: protocol_grant_source(record.source).to_owned(),
-    })
-}
-
-fn next_workspace_grant_counter(grants: &HashMap<WorkspaceGrantId, WorkspaceGrantRecord>) -> u64 {
-    grants
-        .keys()
-        .filter_map(|grant_id| grant_id.as_str().strip_prefix("grant_"))
-        .filter_map(|suffix| suffix.parse::<u64>().ok())
-        .max()
-        .unwrap_or(0)
-        + 1
-}
-
 fn model_invocation_payload(invocation: &ModelInvocation) -> ModelInvocationPayload {
     ModelInvocationPayload {
         requested_model: invocation.requested_model.clone(),
@@ -7162,1072 +5888,6 @@ fn shell_summary(
         "command completed with no output".to_owned()
     } else {
         summary
-    }
-}
-
-fn normalize_voice_checks(checks: Vec<VoiceDoctorCheck>) -> Vec<VoiceDoctorCheck> {
-    checks
-        .into_iter()
-        .filter_map(|check| {
-            let name = check.name.trim();
-            if name.is_empty() {
-                return None;
-            }
-            Some(VoiceDoctorCheck {
-                name: redact(name),
-                status: normalize_voice_check_status(&check.status),
-                message: redact(check.message.trim()),
-            })
-        })
-        .collect()
-}
-
-fn normalize_voice_check_status(status: &str) -> String {
-    match status.trim().to_ascii_lowercase().as_str() {
-        "ok" | "pass" | "passed" | "ready" => "ok",
-        "warn" | "warning" | "degraded" | "unknown" => "warn",
-        "error" | "fail" | "failed" | "blocked" => "error",
-        _ => "warn",
-    }
-    .to_owned()
-}
-
-fn voice_check_summary_status(checks: &[VoiceDoctorCheck]) -> String {
-    if checks.iter().any(|check| check.status == "error") {
-        "error".to_owned()
-    } else if checks.is_empty() || checks.iter().any(|check| check.status == "warn") {
-        "warn".to_owned()
-    } else {
-        "ok".to_owned()
-    }
-}
-
-fn voice_checks_summary(checks: &[VoiceDoctorCheck]) -> String {
-    let errors = checks
-        .iter()
-        .filter(|check| check.status == "error")
-        .count();
-    let warnings = checks.iter().filter(|check| check.status == "warn").count();
-    if errors > 0 {
-        format!("{errors} blocking voice issue{}", plural(errors))
-    } else if warnings > 0 {
-        format!("{warnings} voice warning{}", plural(warnings))
-    } else if checks.is_empty() {
-        "no bridge checks reported".to_owned()
-    } else {
-        "voice bridge ready".to_owned()
-    }
-}
-
-fn plural(count: usize) -> &'static str {
-    if count == 1 {
-        ""
-    } else {
-        "s"
-    }
-}
-
-fn voice_runtime_state(checks: &[VoiceDoctorCheck]) -> VoiceRuntimeState {
-    if checks.iter().any(|check| check.status == "error") {
-        VoiceRuntimeState::Blocked
-    } else if checks.iter().any(|check| check.status == "warn") {
-        VoiceRuntimeState::Degraded
-    } else {
-        VoiceRuntimeState::Ready
-    }
-}
-
-fn normalize_voice_config_string(value: &str) -> String {
-    value.trim().to_owned()
-}
-
-fn is_supported_voice_provider(provider: &str) -> bool {
-    matches!(provider, "edge" | "openai" | "system" | "stub")
-}
-
-fn tool_workspace_summary(input: &serde_json::Value) -> Option<String> {
-    tool_workspace_id(input)
-        .or_else(|| input_string(input, "workspace"))
-        .or_else(|| input_string(input, "cwd"))
-}
-
-fn tool_workspace_id(input: &serde_json::Value) -> Option<String> {
-    input_string(input, "workspace_id")
-}
-
-fn required_tool_access(tool_name: &str) -> WorkspaceAccess {
-    match tool_name {
-        "shell.run" => WorkspaceAccess::Exec,
-        "file.write" | "file.patch" => WorkspaceAccess::Write,
-        _ => WorkspaceAccess::Read,
-    }
-}
-
-fn normalize_workspace_access(access: Vec<WorkspaceAccess>) -> Vec<WorkspaceAccess> {
-    let mut normalized = if access.is_empty() {
-        vec![WorkspaceAccess::Read]
-    } else {
-        access
-    };
-    normalized.sort_by_key(|access| match access {
-        WorkspaceAccess::Read => 0,
-        WorkspaceAccess::Write => 1,
-        WorkspaceAccess::Exec => 2,
-        WorkspaceAccess::Admin => 3,
-    });
-    normalized.dedup();
-    normalized
-}
-
-fn workspace_access_allows(granted: &[WorkspaceAccess], required: WorkspaceAccess) -> bool {
-    granted.contains(&WorkspaceAccess::Admin)
-        || granted.contains(&required)
-        || (required == WorkspaceAccess::Read && granted.contains(&WorkspaceAccess::Write))
-}
-
-fn workspace_grant_matches_agent(
-    grant_agent_id: Option<&AgentId>,
-    request_agent_id: Option<&AgentId>,
-) -> bool {
-    grant_agent_id.is_none() || grant_agent_id == request_agent_id
-}
-
-fn store_workspace_kind(kind: WorkspaceKind) -> StoreWorkspaceKind {
-    match kind {
-        WorkspaceKind::Project => StoreWorkspaceKind::Project,
-        WorkspaceKind::Documents => StoreWorkspaceKind::Documents,
-        WorkspaceKind::Sandbox => StoreWorkspaceKind::Sandbox,
-        WorkspaceKind::Worktree => StoreWorkspaceKind::Worktree,
-    }
-}
-
-fn protocol_workspace_kind(kind: StoreWorkspaceKind) -> WorkspaceKind {
-    match kind {
-        StoreWorkspaceKind::Project => WorkspaceKind::Project,
-        StoreWorkspaceKind::Documents => WorkspaceKind::Documents,
-        StoreWorkspaceKind::Sandbox => WorkspaceKind::Sandbox,
-        StoreWorkspaceKind::Worktree => WorkspaceKind::Worktree,
-    }
-}
-
-fn store_workspace_access(access: WorkspaceAccess) -> StoreWorkspaceAccess {
-    match access {
-        WorkspaceAccess::Read => StoreWorkspaceAccess::Read,
-        WorkspaceAccess::Write => StoreWorkspaceAccess::Write,
-        WorkspaceAccess::Exec => StoreWorkspaceAccess::Exec,
-        WorkspaceAccess::Admin => StoreWorkspaceAccess::Admin,
-    }
-}
-
-fn protocol_workspace_access(access: StoreWorkspaceAccess) -> WorkspaceAccess {
-    match access {
-        StoreWorkspaceAccess::Read => WorkspaceAccess::Read,
-        StoreWorkspaceAccess::Write => WorkspaceAccess::Write,
-        StoreWorkspaceAccess::Exec => WorkspaceAccess::Exec,
-        StoreWorkspaceAccess::Admin => WorkspaceAccess::Admin,
-    }
-}
-
-fn store_workspace_vcs(vcs: Option<&str>) -> WorkspaceVcs {
-    match vcs.unwrap_or_default().trim().to_lowercase().as_str() {
-        "git" => WorkspaceVcs::Git,
-        _ => WorkspaceVcs::None,
-    }
-}
-
-fn store_grant_source(source: &str) -> StoreGrantSource {
-    match source.trim().to_lowercase().as_str() {
-        "route" => StoreGrantSource::Route,
-        "policy" => StoreGrantSource::Policy,
-        "worker_spawn" | "worker-spawn" => StoreGrantSource::WorkerSpawn,
-        _ => StoreGrantSource::User,
-    }
-}
-
-fn protocol_grant_source(source: StoreGrantSource) -> &'static str {
-    match source {
-        StoreGrantSource::Route => "route",
-        StoreGrantSource::User => "user",
-        StoreGrantSource::Policy => "policy",
-        StoreGrantSource::WorkerSpawn => "worker_spawn",
-    }
-}
-
-fn normalize_aliases(aliases: Vec<String>) -> Vec<String> {
-    let mut aliases = aliases
-        .into_iter()
-        .map(|alias| alias.trim().to_owned())
-        .filter(|alias| !alias.is_empty())
-        .collect::<Vec<_>>();
-    aliases.sort();
-    aliases.dedup();
-    aliases
-}
-
-fn canonical_workspace_root(root: &str) -> Result<PathBuf, std::io::Error> {
-    let path = if let Some(rest) = root.strip_prefix("~/") {
-        std::env::var_os("HOME")
-            .map(PathBuf::from)
-            .unwrap_or_else(|| PathBuf::from("~"))
-            .join(rest)
-    } else {
-        PathBuf::from(root)
-    };
-    path.canonicalize()
-}
-
-fn validate_workspace_root(root: &Path, cadis_home: &Path) -> Result<(), ErrorPayload> {
-    if root.parent().is_none() {
-        return Err(tool_error(
-            "workspace_root_too_broad",
-            "workspace root cannot be the filesystem root",
-            false,
-        ));
-    }
-
-    if ["/etc", "/dev", "/proc", "/sys", "/run"]
-        .iter()
-        .any(|path| root == Path::new(path))
-    {
-        return Err(tool_error(
-            "workspace_root_denied",
-            format!(
-                "workspace root {} is a protected system path",
-                root.display()
-            ),
-            false,
-        ));
-    }
-
-    if let Some(home) = std::env::var_os("HOME")
-        .map(PathBuf::from)
-        .and_then(|path| path.canonicalize().ok())
-    {
-        if root == home || home.starts_with(root) {
-            return Err(tool_error(
-                "workspace_root_too_broad",
-                "workspace root cannot be the home directory or an ancestor of it",
-                false,
-            ));
-        }
-
-        for denied in [".ssh", ".aws", ".gnupg", ".config/gh", ".cadis"] {
-            let denied = home.join(denied);
-            if root.starts_with(&denied) {
-                return Err(tool_error(
-                    "workspace_root_denied",
-                    format!(
-                        "workspace root {} is a protected secret path",
-                        root.display()
-                    ),
-                    false,
-                ));
-            }
-        }
-    }
-
-    if let Ok(cadis_home) = cadis_home.canonicalize() {
-        if root == cadis_home || root.starts_with(&cadis_home) || cadis_home.starts_with(root) {
-            return Err(tool_error(
-                "workspace_root_denied",
-                "workspace root cannot be CADIS_HOME or an ancestor/child of it",
-                false,
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-fn validate_shell_cwd(cwd: &Path, cadis_home: &Path) -> Result<(), ErrorPayload> {
-    if cwd.parent().is_none() {
-        return Err(tool_error(
-            "shell_cwd_denied",
-            "shell.run cwd cannot be the filesystem root",
-            false,
-        ));
-    }
-
-    for denied in ["/etc", "/dev", "/proc", "/sys", "/run"] {
-        let denied = Path::new(denied);
-        if cwd == denied || cwd.starts_with(denied) {
-            return Err(tool_error(
-                "shell_cwd_denied",
-                format!("shell.run cwd {} is a protected system path", cwd.display()),
-                false,
-            ));
-        }
-    }
-
-    if let Some(home) = std::env::var_os("HOME")
-        .map(PathBuf::from)
-        .and_then(|path| path.canonicalize().ok())
-    {
-        if cwd == home || home.starts_with(cwd) {
-            return Err(tool_error(
-                "shell_cwd_denied",
-                "shell.run cwd cannot be the home directory or an ancestor of it",
-                false,
-            ));
-        }
-
-        for denied in [".ssh", ".aws", ".gnupg", ".config/gh", ".cadis"] {
-            let denied = home.join(denied);
-            if cwd.starts_with(&denied) {
-                return Err(tool_error(
-                    "shell_cwd_denied",
-                    format!("shell.run cwd {} is a protected secret path", cwd.display()),
-                    false,
-                ));
-            }
-        }
-    }
-
-    if let Ok(cadis_home) = cadis_home.canonicalize() {
-        if cwd == cadis_home || cwd.starts_with(&cadis_home) || cadis_home.starts_with(cwd) {
-            return Err(tool_error(
-                "shell_cwd_denied",
-                "shell.run cwd cannot be CADIS_HOME or an ancestor/child of it",
-                false,
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-fn root_check(name: &str, root: &Path) -> WorkspaceDoctorCheck {
-    if root.is_dir() {
-        WorkspaceDoctorCheck {
-            name: name.to_owned(),
-            status: "ok".to_owned(),
-            message: format!("{} exists", root.display()),
-        }
-    } else {
-        WorkspaceDoctorCheck {
-            name: name.to_owned(),
-            status: "error".to_owned(),
-            message: format!("{} is not a directory", root.display()),
-        }
-    }
-}
-
-fn project_workspace_metadata_checks(workspace: &WorkspaceRecord) -> Vec<WorkspaceDoctorCheck> {
-    if workspace.kind != WorkspaceKind::Project {
-        return Vec::new();
-    }
-
-    let mut checks = project_workspace_metadata_checks_for_root(&workspace.root);
-    checks.extend(project_worker_worktree_checks_for_root(&workspace.root));
-    let Some(metadata) = ProjectWorkspaceStore::new(&workspace.root)
-        .load()
-        .ok()
-        .flatten()
-    else {
-        return checks;
-    };
-
-    if metadata.workspace_id != workspace.id.to_string() {
-        checks.push(WorkspaceDoctorCheck {
-            name: "workspace.metadata.id".to_owned(),
-            status: "error".to_owned(),
-            message: format!(
-                ".cadis/workspace.toml workspace_id '{}' does not match registry id '{}'",
-                metadata.workspace_id, workspace.id
-            ),
-        });
-    }
-
-    let metadata_kind = protocol_workspace_kind(metadata.kind);
-    if metadata_kind != workspace.kind {
-        checks.push(WorkspaceDoctorCheck {
-            name: "workspace.metadata.kind".to_owned(),
-            status: "warn".to_owned(),
-            message: format!(
-                ".cadis/workspace.toml kind {:?} differs from registry kind {:?}",
-                metadata_kind, workspace.kind
-            ),
-        });
-    }
-
-    for (name, path) in [
-        ("workspace.metadata.worktree_root", metadata.worktree_root),
-        ("workspace.metadata.artifact_root", metadata.artifact_root),
-        ("workspace.metadata.media_root", metadata.media_root),
-    ] {
-        if path.is_absolute() {
-            checks.push(WorkspaceDoctorCheck {
-                name: name.to_owned(),
-                status: "warn".to_owned(),
-                message: format!("{} should be project-relative", path.display()),
-            });
-        }
-    }
-
-    checks
-}
-
-fn project_workspace_metadata_checks_for_root(root: &Path) -> Vec<WorkspaceDoctorCheck> {
-    let store = ProjectWorkspaceStore::new(root);
-    match store.load() {
-        Ok(Some(_)) => vec![WorkspaceDoctorCheck {
-            name: "workspace.metadata".to_owned(),
-            status: "ok".to_owned(),
-            message: format!("{} exists", store.workspace_toml_path().display()),
-        }],
-        Ok(None) => vec![WorkspaceDoctorCheck {
-            name: "workspace.metadata".to_owned(),
-            status: "warn".to_owned(),
-            message: format!("{} is missing", store.workspace_toml_path().display()),
-        }],
-        Err(error) => vec![WorkspaceDoctorCheck {
-            name: "workspace.metadata".to_owned(),
-            status: "error".to_owned(),
-            message: format!(
-                "could not read {}: {error}",
-                store.workspace_toml_path().display()
-            ),
-        }],
-    }
-}
-
-fn project_worker_worktree_checks_for_root(root: &Path) -> Vec<WorkspaceDoctorCheck> {
-    let store = ProjectWorkspaceStore::new(root);
-    match store.worker_worktree_diagnostics() {
-        Ok(diagnostics) => diagnostics
-            .into_iter()
-            .map(project_worktree_diagnostic_check)
-            .collect(),
-        Err(error) => vec![WorkspaceDoctorCheck {
-            name: "workspace.worktrees.metadata".to_owned(),
-            status: "error".to_owned(),
-            message: format!("could not inspect worker worktree metadata: {error}"),
-        }],
-    }
-}
-
-fn project_worktree_diagnostic_check(
-    diagnostic: ProjectWorktreeDiagnostic,
-) -> WorkspaceDoctorCheck {
-    WorkspaceDoctorCheck {
-        name: diagnostic.name,
-        status: diagnostic.status,
-        message: diagnostic.message,
-    }
-}
-
-fn tool_command_summary(tool_name: &str, input: &serde_json::Value) -> Option<String> {
-    match tool_name {
-        "shell.run" => input_string(input, "command").or_else(|| {
-            input.get("args").and_then(|value| {
-                value.as_array().map(|items| {
-                    items
-                        .iter()
-                        .filter_map(serde_json::Value::as_str)
-                        .collect::<Vec<_>>()
-                        .join(" ")
-                })
-            })
-        }),
-        "file.patch" => file_patch_path_summary(input),
-        _ => input_string(input, "path"),
-    }
-    .map(|value| redact(&value))
-}
-
-fn file_patch_path_summary(input: &serde_json::Value) -> Option<String> {
-    if let Some(path) = input_string(input, "path").or_else(|| input_string(input, "target")) {
-        return Some(path);
-    }
-
-    input
-        .get("operations")
-        .and_then(serde_json::Value::as_array)
-        .map(|operations| {
-            operations
-                .iter()
-                .filter_map(|operation| {
-                    input_string(operation, "path").or_else(|| input_string(operation, "target"))
-                })
-                .take(FILE_PATCH_OUTPUT_MAX_FILES)
-                .collect::<Vec<_>>()
-                .join(", ")
-        })
-        .filter(|summary| !summary.is_empty())
-}
-
-fn parse_file_patch_operations(
-    input: &serde_json::Value,
-) -> Result<Vec<FilePatchOperation>, ErrorPayload> {
-    let operations = if let Some(operations) = input.get("operations") {
-        let Some(items) = operations.as_array() else {
-            return Err(tool_error(
-                "invalid_tool_input",
-                "file.patch operations must be an array",
-                false,
-            ));
-        };
-        items
-            .iter()
-            .map(parse_file_patch_operation)
-            .collect::<Result<Vec<_>, _>>()?
-    } else {
-        vec![parse_file_patch_operation(input)?]
-    };
-
-    if operations.is_empty() {
-        return Err(tool_error(
-            "invalid_tool_input",
-            "file.patch requires at least one operation",
-            false,
-        ));
-    }
-    if operations.len() > FILE_PATCH_MAX_OPERATIONS {
-        return Err(tool_error(
-            "invalid_tool_input",
-            format!("file.patch supports at most {FILE_PATCH_MAX_OPERATIONS} operations"),
-            false,
-        ));
-    }
-
-    Ok(operations)
-}
-
-fn parse_file_patch_operation(
-    input: &serde_json::Value,
-) -> Result<FilePatchOperation, ErrorPayload> {
-    let path = input_string(input, "path")
-        .or_else(|| input_string(input, "target"))
-        .ok_or_else(|| tool_error("invalid_tool_input", "file.patch requires path", false))?;
-    let action = input_string(input, "op")
-        .or_else(|| input_string(input, "action"))
-        .map(|value| value.to_ascii_lowercase());
-
-    match action.as_deref() {
-        Some("write") | Some("replace_file") | Some("create") => {
-            let content = input_raw_string(input, "content").ok_or_else(|| {
-                tool_error(
-                    "invalid_tool_input",
-                    "file.patch write operation requires content",
-                    false,
-                )
-            })?;
-            validate_patch_text_size("content", &content)?;
-            Ok(FilePatchOperation::Write { path, content })
-        }
-        Some("replace") => parse_file_patch_replace(path, input),
-        Some(other) => Err(tool_error(
-            "invalid_tool_input",
-            format!("unsupported file.patch operation '{other}'"),
-            false,
-        )),
-        None if input.get("content").is_some() => {
-            let content = input_raw_string(input, "content").ok_or_else(|| {
-                tool_error(
-                    "invalid_tool_input",
-                    "file.patch content must be a string",
-                    false,
-                )
-            })?;
-            validate_patch_text_size("content", &content)?;
-            Ok(FilePatchOperation::Write { path, content })
-        }
-        None => parse_file_patch_replace(path, input),
-    }
-}
-
-fn parse_file_patch_replace(
-    path: String,
-    input: &serde_json::Value,
-) -> Result<FilePatchOperation, ErrorPayload> {
-    let old = input_raw_string(input, "old").ok_or_else(|| {
-        tool_error(
-            "invalid_tool_input",
-            "file.patch replace operation requires old",
-            false,
-        )
-    })?;
-    let new = input_raw_string(input, "new").ok_or_else(|| {
-        tool_error(
-            "invalid_tool_input",
-            "file.patch replace operation requires new",
-            false,
-        )
-    })?;
-    if old.is_empty() {
-        return Err(tool_error(
-            "invalid_tool_input",
-            "file.patch replace operation old cannot be empty",
-            false,
-        ));
-    }
-    validate_patch_text_size("old", &old)?;
-    validate_patch_text_size("new", &new)?;
-    Ok(FilePatchOperation::Replace { path, old, new })
-}
-
-fn validate_patch_text_size(label: &str, value: &str) -> Result<(), ErrorPayload> {
-    if value.len() > FILE_PATCH_MAX_FILE_BYTES {
-        Err(tool_error(
-            "file_patch_too_large",
-            format!("file.patch {label} exceeds {FILE_PATCH_MAX_FILE_BYTES} bytes"),
-            false,
-        ))
-    } else {
-        Ok(())
-    }
-}
-
-fn validate_file_patch_input(
-    workspace: &Path,
-    input: &serde_json::Value,
-) -> Result<(), ErrorPayload> {
-    let operations = parse_file_patch_operations(input)?;
-    for operation in &operations {
-        let target = resolve_file_patch_target(workspace, operation.path())?;
-        validate_file_patch_target(operation, &target)?;
-    }
-    Ok(())
-}
-
-fn validate_file_patch_target(
-    operation: &FilePatchOperation,
-    target: &Path,
-) -> Result<(), ErrorPayload> {
-    match operation {
-        FilePatchOperation::Write { .. } => {
-            if let Ok(metadata) = fs::metadata(target) {
-                if !metadata.is_file() {
-                    return Err(tool_error(
-                        "unsupported_file_type",
-                        "file.patch can only write regular files",
-                        false,
-                    ));
-                }
-            }
-            Ok(())
-        }
-        FilePatchOperation::Replace { .. } => {
-            let metadata = fs::metadata(target).map_err(|error| {
-                tool_error(
-                    "path_resolution_failed",
-                    format!("could not read file.patch target: {error}"),
-                    false,
-                )
-            })?;
-            if !metadata.is_file() {
-                return Err(tool_error(
-                    "unsupported_file_type",
-                    "file.patch can only replace regular files",
-                    false,
-                ));
-            }
-            if metadata.len() > FILE_PATCH_MAX_FILE_BYTES as u64 {
-                return Err(tool_error(
-                    "file_patch_too_large",
-                    format!("file.patch target exceeds {FILE_PATCH_MAX_FILE_BYTES} bytes"),
-                    false,
-                ));
-            }
-            Ok(())
-        }
-    }
-}
-
-fn prepare_file_patch(
-    workspace: &Path,
-    operations: &[FilePatchOperation],
-) -> Result<Vec<PreparedFilePatch>, ErrorPayload> {
-    let mut staged = HashMap::<PathBuf, String>::new();
-    let mut prepared = Vec::new();
-
-    for operation in operations {
-        let path = resolve_file_patch_target(workspace, operation.path())?;
-        validate_file_patch_target(operation, &path)?;
-        let mtime = fs::metadata(&path).ok().and_then(|m| m.modified().ok());
-        let content = match operation {
-            FilePatchOperation::Write { content, .. } => content.clone(),
-            FilePatchOperation::Replace { old, new, .. } => {
-                let current = match staged.get(&path) {
-                    Some(content) => content.clone(),
-                    None => read_patch_target(&path)?,
-                };
-                replace_once(&current, old, new)?
-            }
-        };
-        validate_patch_text_size("result", &content)?;
-        staged.insert(path.clone(), content.clone());
-        prepared.push(PreparedFilePatch {
-            display_path: display_relative_path(workspace, &path),
-            path,
-            action: operation.action(),
-            content,
-            mtime,
-        });
-    }
-
-    Ok(prepared)
-}
-
-fn read_patch_target(path: &Path) -> Result<String, ErrorPayload> {
-    fs::read_to_string(path).map_err(|error| {
-        tool_error(
-            "file_patch_read_failed",
-            format!("could not read patch target: {error}"),
-            false,
-        )
-    })
-}
-
-fn replace_once(content: &str, old: &str, new: &str) -> Result<String, ErrorPayload> {
-    let matches = content.match_indices(old).take(2).count();
-    match matches {
-        0 => Err(tool_error(
-            "file_patch_replace_mismatch",
-            "file.patch old text was not found exactly once",
-            false,
-        )),
-        1 => Ok(content.replacen(old, new, 1)),
-        _ => Err(tool_error(
-            "file_patch_replace_ambiguous",
-            "file.patch old text matched more than once",
-            false,
-        )),
-    }
-}
-
-fn resolve_file_patch_target(workspace: &Path, user_path: &str) -> Result<PathBuf, ErrorPayload> {
-    let relative = Path::new(user_path);
-    if relative.is_absolute() || path_has_parent_or_root(relative) {
-        return Err(tool_error(
-            "outside_workspace",
-            "file.patch paths must be relative to the workspace",
-            false,
-        ));
-    }
-    if relative.file_name().is_none() {
-        return Err(tool_error(
-            "invalid_tool_input",
-            "file.patch path must name a file",
-            false,
-        ));
-    }
-    validate_file_patch_relative_path(relative)?;
-
-    let workspace = workspace.canonicalize().map_err(|error| {
-        tool_error(
-            "path_resolution_failed",
-            format!("could not resolve workspace: {error}"),
-            false,
-        )
-    })?;
-    let candidate = workspace.join(relative);
-    if let Ok(resolved) = candidate.canonicalize() {
-        if !resolved.starts_with(&workspace) {
-            return Err(tool_error(
-                "outside_workspace",
-                "file.patch target resolves outside the workspace",
-                false,
-            ));
-        }
-        if resolved.is_dir() {
-            return Err(tool_error(
-                "unsupported_file_type",
-                "file.patch target must be a file",
-                false,
-            ));
-        }
-        if let Ok(resolved_relative) = resolved.strip_prefix(&workspace) {
-            validate_file_patch_relative_path(resolved_relative)?;
-        }
-        return Ok(resolved);
-    }
-
-    let parent = candidate.parent().ok_or_else(|| {
-        tool_error(
-            "invalid_tool_input",
-            "file.patch path must have a parent directory",
-            false,
-        )
-    })?;
-    let parent = parent.canonicalize().map_err(|error| {
-        tool_error(
-            "path_resolution_failed",
-            format!("could not resolve file.patch parent: {error}"),
-            false,
-        )
-    })?;
-    if !parent.starts_with(&workspace) {
-        return Err(tool_error(
-            "outside_workspace",
-            "file.patch parent resolves outside the workspace",
-            false,
-        ));
-    }
-    let file_name = candidate.file_name().ok_or_else(|| {
-        tool_error(
-            "invalid_tool_input",
-            "file.patch path must name a file",
-            false,
-        )
-    })?;
-    let resolved = parent.join(file_name);
-    if let Ok(resolved_relative) = resolved.strip_prefix(&workspace) {
-        validate_file_patch_relative_path(resolved_relative)?;
-    }
-    Ok(resolved)
-}
-
-fn validate_file_patch_relative_path(path: &Path) -> Result<(), ErrorPayload> {
-    if file_patch_path_is_protected(path) {
-        return Err(tool_error(
-            "protected_path",
-            "file.patch refuses to modify protected workspace metadata paths",
-            false,
-        ));
-    }
-    if file_patch_path_is_secret_like(path) {
-        return Err(tool_error(
-            "secret_path_rejected",
-            "file.patch refuses to modify secret-like paths",
-            false,
-        ));
-    }
-    Ok(())
-}
-
-fn path_has_parent_or_root(path: &Path) -> bool {
-    path.components().any(|component| {
-        matches!(
-            component,
-            Component::ParentDir | Component::RootDir | Component::Prefix(_)
-        )
-    })
-}
-
-fn file_patch_path_is_protected(path: &Path) -> bool {
-    path.components().any(|component| match component {
-        Component::Normal(value) => matches!(value.to_str(), Some(".git" | ".cadis")),
-        _ => false,
-    })
-}
-
-fn file_patch_path_is_secret_like(path: &Path) -> bool {
-    path.components().any(|component| {
-        let Component::Normal(value) = component else {
-            return false;
-        };
-        let name = value.to_string_lossy().to_ascii_lowercase();
-        name == ".env"
-            || name.starts_with(".env.")
-            || matches!(
-                name.as_str(),
-                ".netrc"
-                    | ".npmrc"
-                    | ".pypirc"
-                    | ".git-credentials"
-                    | "id_rsa"
-                    | "id_dsa"
-                    | "id_ecdsa"
-                    | "id_ed25519"
-                    | ".ssh"
-                    | ".aws"
-                    | ".gnupg"
-            )
-            || name.ends_with(".pem")
-            || name.ends_with(".key")
-            || name.ends_with(".p12")
-            || name.ends_with(".pfx")
-            || name.contains("secret")
-            || name.contains("credential")
-            || name.contains("token")
-            || name.contains("api_key")
-            || name.contains("apikey")
-            || name.contains("private_key")
-    })
-}
-
-/// Item 17: Returns true if an environment variable name looks secret-bearing.
-#[allow(dead_code)]
-fn is_secret_env_var(name: &str) -> bool {
-    let lower = name.to_ascii_lowercase();
-    lower.contains("secret")
-        || lower.contains("password")
-        || lower.contains("token")
-        || lower.contains("api_key")
-        || lower.contains("apikey")
-        || lower.contains("private_key")
-        || lower.contains("credential")
-        || lower.ends_with("_key")
-        || matches!(
-            lower.as_str(),
-            "aws_secret_access_key"
-                | "openai_api_key"
-                | "cadis_openai_api_key"
-                | "ssh_auth_sock"
-                | "github_token"
-                | "gh_token"
-        )
-}
-
-/// Item 17: Returns true if a config value looks like it contains a secret.
-#[allow(dead_code)]
-fn is_secret_config_value(value: &str) -> bool {
-    let lower = value.to_ascii_lowercase();
-    lower.contains("secret=")
-        || lower.contains("password=")
-        || lower.contains("token=")
-        || lower.contains("api_key=")
-        || lower.contains("apikey=")
-        || lower.contains("private_key=")
-}
-
-fn validate_git_pathspec(pathspec: &str) -> Result<String, ErrorPayload> {
-    let trimmed = pathspec.trim();
-    if trimmed.is_empty() {
-        return Err(tool_error(
-            "invalid_tool_input",
-            "git.diff pathspec cannot be empty",
-            false,
-        ));
-    }
-    if trimmed.starts_with(':') {
-        return Err(tool_error(
-            "invalid_tool_input",
-            "git.diff pathspec magic is not supported",
-            false,
-        ));
-    }
-
-    let path = Path::new(trimmed);
-    if path.is_absolute()
-        || path.components().any(|component| {
-            matches!(
-                component,
-                Component::ParentDir | Component::RootDir | Component::Prefix(_)
-            )
-        })
-    {
-        return Err(tool_error(
-            "outside_workspace",
-            "git.diff pathspec must be relative to the workspace",
-            false,
-        ));
-    }
-
-    Ok(trimmed.to_owned())
-}
-
-fn resolve_inside_workspace(workspace: &Path, user_path: &str) -> Result<PathBuf, ErrorPayload> {
-    let candidate = PathBuf::from(user_path);
-    let candidate = if candidate.is_absolute() {
-        candidate
-    } else {
-        workspace.join(candidate)
-    };
-    let resolved = candidate.canonicalize().map_err(|error| {
-        tool_error(
-            "path_resolution_failed",
-            format!("could not resolve {}: {error}", candidate.display()),
-            false,
-        )
-    })?;
-
-    if resolved.starts_with(workspace) {
-        Ok(resolved)
-    } else {
-        Err(tool_error(
-            "outside_workspace",
-            format!(
-                "{} is outside workspace {}",
-                resolved.display(),
-                workspace.display()
-            ),
-            false,
-        ))
-    }
-}
-
-fn display_relative_path(workspace: &Path, path: &Path) -> String {
-    path.strip_prefix(workspace)
-        .unwrap_or(path)
-        .display()
-        .to_string()
-}
-
-fn search_files(
-    workspace: &Path,
-    root: &Path,
-    query: &str,
-    max_results: usize,
-    matches: &mut Vec<SearchMatch>,
-) {
-    if matches.len() >= max_results {
-        return;
-    }
-    let Ok(entries) = fs::read_dir(root) else {
-        return;
-    };
-
-    for entry in entries.flatten() {
-        if matches.len() >= max_results {
-            return;
-        }
-        let path = entry.path();
-        let name = entry.file_name();
-        if name == ".git" || name == "target" {
-            continue;
-        }
-        let Ok(resolved) = path.canonicalize() else {
-            continue;
-        };
-        if !resolved.starts_with(workspace) {
-            continue;
-        }
-        let Ok(metadata) = fs::metadata(&resolved) else {
-            continue;
-        };
-        if metadata.is_dir() {
-            search_files(workspace, &resolved, query, max_results, matches);
-        } else if metadata.is_file() && metadata.len() <= FILE_SEARCH_LIMIT_BYTES {
-            search_file(workspace, &resolved, query, max_results, matches);
-        }
-    }
-}
-
-fn search_file(
-    workspace: &Path,
-    path: &Path,
-    query: &str,
-    max_results: usize,
-    matches: &mut Vec<SearchMatch>,
-) {
-    let Ok(content) = fs::read_to_string(path) else {
-        return;
-    };
-    for (index, line) in content.lines().enumerate() {
-        if matches.len() >= max_results {
-            return;
-        }
-        if line.contains(query) {
-            matches.push(SearchMatch {
-                path: display_relative_path(workspace, path),
-                line_number: index + 1,
-                line: line.to_owned(),
-            });
-        }
-    }
-}
-
-fn tool_error(
-    code: impl Into<String>,
-    message: impl Into<String>,
-    retryable: bool,
-) -> ErrorPayload {
-    ErrorPayload {
-        code: code.into(),
-        message: redact(&message.into()),
-        retryable,
     }
 }
 
@@ -8407,17 +6067,6 @@ fn default_agent_specialist_profile(
     }
 }
 
-fn normalize_role(value: &str) -> String {
-    let collapsed: String = value.split_whitespace().collect::<Vec<_>>().join(" ");
-    if collapsed.is_empty() {
-        return collapsed;
-    }
-    collapsed
-        .parse::<cadis_protocol::AgentRole>()
-        .map(|r| r.as_str().to_owned())
-        .unwrap_or(collapsed)
-}
-
 fn default_agent_name(role: &str, agent_id: &AgentId) -> String {
     let name = role
         .split(|character: char| !character.is_ascii_alphanumeric())
@@ -8473,163 +6122,6 @@ fn agent_session_status_label(status: AgentSessionStatus) -> &'static str {
         AgentSessionStatus::TimedOut => "timed_out",
         AgentSessionStatus::BudgetExceeded => "budget_exceeded",
     }
-}
-
-fn leading_mention(content: &str) -> Option<(String, String)> {
-    let trimmed = content.trim_start();
-    let rest = trimmed.strip_prefix('@')?;
-    let mut end = 0;
-    for (index, character) in rest.char_indices() {
-        if character.is_ascii_alphanumeric() || character == '_' || character == '-' {
-            end = index + character.len_utf8();
-        } else {
-            break;
-        }
-    }
-    if end == 0 {
-        return None;
-    }
-    let mention = rest[..end].to_owned();
-    let remaining = rest[end..].trim_start().to_owned();
-    Some((mention, remaining))
-}
-
-fn parse_orchestrator_action(
-    content: &str,
-    config: &OrchestratorConfig,
-) -> Option<ExplicitOrchestratorAction> {
-    let trimmed = content.trim_start();
-    let (command, rest) = trimmed
-        .split_once(char::is_whitespace)
-        .unwrap_or((trimmed, ""));
-    let command = command.strip_prefix('/')?.to_ascii_lowercase();
-    let rest = rest.trim_start();
-
-    match command.as_str() {
-        "route" | "delegate" => {
-            let (mention, remaining) = leading_mention(rest)?;
-            let content = normalize_route_content(remaining);
-            let worker_summary = format!("Route @{mention}: {}", summarize_task(&content));
-            Some(ExplicitOrchestratorAction::Route {
-                mention,
-                content,
-                worker_summary,
-            })
-        }
-        "spawn" => {
-            let (role, content) = parse_spawn_action(rest, true, &config.default_worker_role)?;
-            let content = normalize_route_content(content);
-            let worker_summary = format!("Spawn {role}: {}", summarize_task(&content));
-            Some(ExplicitOrchestratorAction::Spawn {
-                role,
-                content,
-                worker_summary,
-            })
-        }
-        "worker" => {
-            let (role, content) = parse_spawn_action(rest, false, &config.default_worker_role)?;
-            let content = normalize_route_content(content);
-            let worker_summary = format!("Worker {role}: {}", summarize_task(&content));
-            Some(ExplicitOrchestratorAction::Spawn {
-                role,
-                content,
-                worker_summary,
-            })
-        }
-        _ => None,
-    }
-}
-
-fn parse_spawn_action(
-    rest: &str,
-    require_role: bool,
-    default_worker_role: &str,
-) -> Option<(String, String)> {
-    let rest = rest.trim();
-    if rest.is_empty() {
-        if require_role {
-            return None;
-        }
-        return Some((default_worker_role.to_owned(), String::new()));
-    }
-
-    if let Some((role, content)) = rest.split_once(':') {
-        let role = normalize_role(role);
-        if role.is_empty() {
-            return None;
-        }
-        return Some((role, content.trim_start().to_owned()));
-    }
-
-    if require_role {
-        Some((normalize_role(rest), String::new()))
-    } else {
-        Some((default_worker_role.to_owned(), rest.to_owned()))
-    }
-}
-
-fn normalize_route_content(content: String) -> String {
-    let content = content.trim().to_owned();
-    if content.is_empty() {
-        "Continue.".to_owned()
-    } else {
-        content
-    }
-}
-
-fn summarize_task(content: &str) -> String {
-    let summary = content
-        .split_whitespace()
-        .take(12)
-        .collect::<Vec<_>>()
-        .join(" ");
-    if summary.is_empty() {
-        "Continue.".to_owned()
-    } else {
-        summary
-    }
-}
-
-/// Parsed model-driven spawn directive from a model response.
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ModelSpawnDirective {
-    role: String,
-    task: String,
-}
-
-/// Maximum spawn directives parsed from a single model response.
-const MAX_SPAWN_DIRECTIVES_PER_RESPONSE: usize = 3;
-/// Maximum length for a model-driven spawn role.
-const SPAWN_ROLE_MAX_LEN: usize = 64;
-/// Maximum length for a model-driven spawn task.
-const SPAWN_TASK_MAX_LEN: usize = 256;
-
-/// Parses `[SPAWN role: task]` directives from model response text.
-fn parse_model_spawn_directives(content: &str) -> Vec<ModelSpawnDirective> {
-    let mut directives = Vec::new();
-    for line in content.lines() {
-        if directives.len() >= MAX_SPAWN_DIRECTIVES_PER_RESPONSE {
-            break;
-        }
-        let trimmed = line.trim();
-        if let Some(inner) = trimmed
-            .strip_prefix("[SPAWN ")
-            .and_then(|s| s.strip_suffix(']'))
-        {
-            if let Some((role, task)) = inner.split_once(':') {
-                let role = truncate_to_utf8_boundary(&normalize_role(role), SPAWN_ROLE_MAX_LEN)
-                    .0
-                    .to_owned();
-                let task = truncate_to_utf8_boundary(task.trim(), SPAWN_TASK_MAX_LEN)
-                    .0
-                    .to_owned();
-                if !role.is_empty() && !task.is_empty() {
-                    directives.push(ModelSpawnDirective { role, task });
-                }
-            }
-        }
-    }
-    directives
 }
 
 fn planned_worker_worktree(
@@ -9437,10 +6929,6 @@ fn json_u32(value: &serde_json::Value, key: &str) -> Option<u32> {
 
 fn json_i64(value: &serde_json::Value, key: &str) -> Option<i64> {
     value.get(key).and_then(serde_json::Value::as_i64)
-}
-
-fn clamp_voice_adjustment(value: i16) -> i16 {
-    value.clamp(-50, 50)
 }
 
 fn merge_json(mut base: serde_json::Value, patch: serde_json::Value) -> serde_json::Value {
