+++ docs/progress.json
@@ -4,7 +4,7 @@
     "total": 404,
     "percent": 100
   },
-  "milestone": "v0.9.2-rc",
+  "milestone": "v1.0-stable",
   "next_milestone": "v1.0-stable",
   "next_target_percent": 100,
   "updated_at": "2026-04-29T03:24:00Z"
